requirements.txt
asgiref==3.9.2
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==5.2.7
django-classy-tags==4.1.0
django-cms==5.0.4
django-entangled==0.6.3
django-filer==3.3.2
django-formtools==2.5.1
django-fsm==2.8.2
django-parler==2.3
django-polymorphic==4.1.0
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==3.0.1
djangocms-attributes-field==4.1.1
djangocms-frontend==2.2.0
djangocms-link==5.1.0
djangocms-simple-admin-style==1.1.5
djangocms-text==0.9.3
djangocms-versioning==2.4.0
easy-thumbnails==2.10.1
freetype-py==2.5.1
lxml==6.0.2
nh3==0.3.1
packaging==25.0
pillow==11.3.0
pycairo==1.28.0
reportlab==4.4.4
rlPyCairo==0.4.0
setuptools==80.9.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
webencodings==0.5.1
